Mengaktifkan kembali jwt di beberapa endpoint

+++ src/middlewares/photoLimit.js
@@ -3,16 +3,18 @@ const deleteFile = require('../utils/deleteFile');
 
 module.exports = (req, res, next) => {
   try {
-    if (req.files.photo) {
-      if (req.files.photo[0].size > 2000000) {
-        deleteFile(req.files.photo[0].path);
+    if (req.files) {
+      if (req.files.photo) {
+        if (req.files.photo[0].size > 2000000) {
+          deleteFile(req.files.photo[0].path);
 
-        failed(res, {
-          code: 400,
-          payload: 'File photo too large, max 2mb',
-          message: 'Upload File Error',
-        });
-        return;
+          failed(res, {
+            code: 400,
+            payload: 'File photo too large, max 2mb',
+            message: 'Upload File Error',
+          });
+          return;
+        }
       }
     }
